Global Dialogue (Event-Based)
This was the most annoying bug ever. It took me like 2 hours to figure out. Obviously, the "auto destroy" dialogues, hide before getting destroyed, but even if you don't touch them (?) So they'd proc a "hide dialogue" in the middle of your regular dialogue. The fix was to just delay the Haneko ending dialogue, so it doesn't get hidden by the automatic-destruction dialogues caused by the attacks.